Fixed shade to string method and reduced solution complexity
